+++ packages/wallet-management/src/connectors/walletConnect.ts
@@ -1,7 +1,7 @@
 import type { WalletConnectParameters } from '@wagmi/connectors';
 import { walletConnect as _walletConnect } from '@wagmi/connectors';
 
-export const createWalletConnectConnector = /*@__PURE__*/ (
+export const createWalletConnectConnector = /*#__PURE__*/ (
   params: WalletConnectParameters,
 ) =>
   _walletConnect({