+++ packages/widget/src/pages/TransactionDetailsPage/TransactionDetailsPage.tsx
@@ -1,10 +1,11 @@
 import type { FullStatusData } from '@lifi/sdk';
 import { ContentCopyRounded } from '@mui/icons-material';
-import { Box, IconButton, Typography } from '@mui/material';
+import { Box, Typography } from '@mui/material';
 import { useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useLocation } from 'react-router-dom';
 import { Card } from '../../components/Card/Card.js';
+import { CardIconButton } from '../../components/Card/CardIconButton.js';
 import { CardTitle } from '../../components/Card/CardTitle.js';
 import { ContractComponent } from '../../components/ContractComponent/ContractComponent.js';
 import { PageContainer } from '../../components/PageContainer.js';
@@ -89,66 +90,63 @@ export const TransactionDetailsPage: React.FC = () => {
       bottomGutters
       sx={mobileLayout ? { justifyContent: 'space-between' } : {}}
     >
-      <Box>
+      <Box
+        sx={{
+          display: 'flex',
+          flex: 1,
+          justifyContent: 'space-between',
+        }}
+        pb={1}
+      >
+        <Typography fontSize={12}>
+          {startedAt.toLocaleString(i18n.language, {
+            dateStyle: 'long',
+          })}
+        </Typography>
+        <Typography fontSize={12}>
+          {startedAt.toLocaleString(i18n.language, {
+            timeStyle: 'short',
+          })}
+        </Typography>
+      </Box>
+      {getStepList(routeExecution?.route, subvariant)}
+      {subvariant === 'custom' && contractSecondaryComponent ? (
+        <ContractComponent sx={{ marginTop: 2 }}>
+          {contractSecondaryComponent}
+        </ContractComponent>
+      ) : null}
+      {routeExecution?.route ? (
+        <TransactionDetails
+          route={routeExecution?.route}
+          sx={{ marginTop: 2 }}
+        />
+      ) : null}
+      <Card sx={{ marginTop: 2 }}>
         <Box
           sx={{
             display: 'flex',
-            justifyContent: 'space-between',
+            flex: 1,
           }}
-          pb={1}
         >
-          <Typography fontSize={12}>
-            {startedAt.toLocaleString(i18n.language, {
-              dateStyle: 'long',
-            })}
-          </Typography>
-          <Typography fontSize={12}>
-            {startedAt.toLocaleString(i18n.language, {
-              timeStyle: 'short',
-            })}
-          </Typography>
-        </Box>
-        {getStepList(routeExecution?.route, subvariant)}
-        {subvariant === 'custom' && contractSecondaryComponent ? (
-          <ContractComponent sx={{ marginTop: 2 }}>
-            {contractSecondaryComponent}
-          </ContractComponent>
-        ) : null}
-        {routeExecution?.route ? (
-          <TransactionDetails
-            route={routeExecution?.route}
-            sx={{ marginTop: 2 }}
-          />
-        ) : null}
-      </Box>
-      <Box>
-        <Card sx={{ marginTop: 2 }}>
-          <Box
-            sx={{
-              display: 'flex',
-              flex: 1,
-            }}
-          >
-            <CardTitle flex={1}>{t('main.supportId')}</CardTitle>
-            <Box mr={1} mt={1}>
-              <IconButton size="medium" onClick={copySupportId}>
-                <ContentCopyRounded fontSize="small" />
-              </IconButton>
-            </Box>
+          <CardTitle flex={1}>{t('main.supportId')}</CardTitle>
+          <Box mr={2} mt={1}>
+            <CardIconButton size="small" onClick={copySupportId}>
+              <ContentCopyRounded fontSize="inherit" />
+            </CardIconButton>
           </Box>
-          <Typography
-            variant="body2"
-            pt={1}
-            pb={2}
-            px={2}
-            sx={{ wordBreak: 'break-all' }}
-          >
-            {supportId}
-          </Typography>
-        </Card>
-        <Box mt={2}>
-          <ContactSupportButton supportId={supportId} />
         </Box>
+        <Typography
+          variant="body2"
+          pt={1}
+          pb={2}
+          px={2}
+          sx={{ wordBreak: 'break-all' }}
+        >
+          {supportId}
+        </Typography>
+      </Card>
+      <Box mt={2}>
+        <ContactSupportButton supportId={supportId} />
       </Box>
     </PageContainer>
   );